Feature/6396 inspector (#6568)
* Moved more of the schemaEditor logic from ./shared into the SchemaEditor in order to place Schema inspector next to the editor and toolbar.
Refactored some imports.

* Updated failing tests and removed redundant duplicate testing.

* Simplified save-logic by moving the save button to the SchemaEditorApp
