+++ src/studio/src/designer/frontend/packages/schema-editor/src/index.ts
@@ -1,3 +1,3 @@
-import SchemaEditorApp from './SchemaEditorApp';
+import SchemaEditorApp, { IAppProps as ISchemaEditorProps } from './SchemaEditorApp';
 
-export { SchemaEditorApp };
+export { SchemaEditorApp as SchemaEditor, ISchemaEditorProps };
