+++ src/studio/src/designer/frontend/packages/schema-editor/src/types/index.ts
@@ -9,7 +9,7 @@ export interface ISchemaState {
   navigate?: string; // used to trigger navigation in tree, the value is not used.
 }
 export interface ILanguage {
-  [key: string]: string;
+  [key: string]: string | ILanguage;
 }
 export interface ISetValueAction {
   path: string,
