+++ src/studio/src/designer/frontend/shared/package.json
@@ -44,6 +44,7 @@
     "@types/react-html-parser": "~2.0.2",
     "@types/react-router-dom": "~5.1.8",
     "@types/react-test-renderer": "~17.0.1",
+    "@types/redux-mock-store": "~1.0.3",
     "altinn-designsystem": "2.7.12",
     "babel-jest": "~27.0.6",
     "babel-loader": "~8.2.2",
@@ -63,6 +64,7 @@
     "react-dom": "~17.0.2",
     "react-styleguidist": "~11.1.7",
     "react-test-renderer": "~17.0.2",
+    "redux-mock-store": "~1.5.4",
     "ts-jest": "~27.0.4"
   },
   "peerDependencies": {
@@ -79,7 +81,8 @@
     ],
     "moduleNameMapper": {
       "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/test/__mocks__/fileMock.js",
-      "\\.(css|less)$": "<rootDir>/test/__mocks__/styleMock.js"
+      "\\.(css|less)$": "<rootDir>/test/__mocks__/styleMock.js",
+      "^@altinn/schema-editor/(.*)": "<rootDir>/../packages/schema-editor/src/$1"
     },
     "testRegex": "(/test/__tests__/.*|\\.(test|spec))\\.(ts|tsx|js)$",
     "moduleFileExtensions": [
